requirements.txt
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
appdirs==1.4.4
arcgis==1.8.1
argh==0.26.2
asgiref==3.7.2
asn1crypto==1.3.0
astroid==2.4.1
astropy==4.0.1.post1
atomicwrites==1.4.0
attrs==23.1.0
autopep8==1.4.4
Babel==2.12.1
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bar-chart-race==0.1.0
basemap==1.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.4
blinker==1.4
bokeh==2.0.1
boto==2.49.0
Bottleneck==1.3.2
branca==0.4.1
Cartopy==0.17.0
certifi==2020.4.5.2
cffi==1.14.0
cftime==1.1.3
chardet==3.0.4
click==8.1.4
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.4.1
clyent==1.2.2
colorama==0.4.6
colorcet==2.0.2
comtypes==1.1.7
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
ConfigUpdater==3.1.1
contextlib2==0.6.0.post1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.17.2
datashader==0.10.0
datashape==0.5.4
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.14
descartes==1.1.0
diff-match-patch==20181111
distlib==0.3.7
distributed==2.17.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.20
django-cors-headers==4.1.0
django-geojson==3.0.0
django-leaflet==0.26.0
django-toolbelt==0.0.1
djangorestframework==3.14.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
ffmpeg-python==0.2.0
filelock==3.12.2
Fiona==1.8.4
flake8==3.7.9
Flask==2.2.5
Flask-Babel==2.0.0
Flask-Limiter==3.3.1
Flask-WTF==1.1.1
folium==0.11.0
fsspec==0.7.4
future==0.18.2
GDAL==2.4.1
geographiclib==1.50
geojson==2.5.0
geopandas==0.7.0
geopy==1.22.0
geos==0.2.2
geoviews==1.8.1
gevent==1.4.0
glob2==0.7
googlemaps==4.4.1
greenlet==2.0.2
gunicorn==20.0.4
h5py==2.10.0
HeapDict==1.0.1
holoviews==1.13.2
html5lib==1.0.1
hypothesis==5.11.0
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==4.13.0
importlib-resources==5.12.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==2.1.2
jdcal==1.4.1
jedi==0.15.2
Jinja2==3.1.2
joblib==0.15.1
json5==0.9.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.4
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lerc==0.1.0
libarchive-c==2.9
limits==3.5.0
llvmlite==0.32.1
locket==0.2.0
lxml==4.5.0
Markdown==3.2.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.2.1
mccabe==0.6.1
mdurl==0.1.2
menuinst==1.4.16
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.3.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
munch==2.5.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.6
netCDF4==1.4.2
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
ntlm-auth==1.4.0
numba==0.49.1
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.0.1
olefile==0.46
openpyxl==3.0.3
ordered-set==4.1.0
osmnx==0.14.0
OWSLib==0.19.2
packaging==21.3
pandas==1.0.3
pandocfilters==1.4.2
panel==0.9.5
param==1.9.3
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2023.7.23
pkginfo==1.5.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
plotly==4.8.1
plotly-express==0.4.1
pluggy==0.13.1
ply==3.11
prison==0.2.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psutil==5.7.0
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pyct==0.4.6
pycurl==7.43.0.5
pydocstyle==4.0.1
pyepsg==0.4.0
pyflakes==2.1.1
Pygments==2.15.1
PyJWT==2.7.0
pykdtree==1.3.1
pylint==2.5.2
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
PyNaCl==1.3.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==3.1.0
pyproj==2.6.1.post1
pyreadline==2.1
pyrsistent==0.19.3
pyshp==2.1.0
PySocks==1.7.1
pytest==5.4.2
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.2
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2023.3
pyviz-comms==0.7.5
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
retrying==1.3.3
rich==13.4.2
rope==0.17.0
Rtree==0.9.4
ruamel-yaml==0.15.87
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.7.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.16.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.4.49
sqlparse==0.3.1
static3==0.7.0
statsmodels==0.11.1
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==4.7.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.8
virtualenv==20.24.3
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.6
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
winkerberos==0.7.0
wrapt==1.15.0
WTForms==3.0.1
xarray==0.15.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.19.4
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
zict==2.0.0
zipp==3.15.0
